# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/restaurants/{restaurantId}/fulfillment/orders/{orderId}/reject" : {
      "post" : {
        "tags" : [ "fulfillment-controller" ],
        "operationId" : "rejectOrder",
        "parameters" : [ {
          "name" : "restaurantId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/restaurants/{restaurantId}/fulfillment/orders/{orderId}/refund/partial" : {
      "post" : {
        "tags" : [ "fulfillment-controller" ],
        "operationId" : "partialRefund",
        "parameters" : [ {
          "name" : "restaurantId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PartialRefundRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/restaurants/{restaurantId}/fulfillment/orders/{orderId}/prepare" : {
      "post" : {
        "tags" : [ "fulfillment-controller" ],
        "operationId" : "prepareOrder",
        "parameters" : [ {
          "name" : "restaurantId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/restaurants/{restaurantId}/fulfillment/orders/{orderId}/cancel" : {
      "post" : {
        "tags" : [ "fulfillment-controller" ],
        "operationId" : "cancelOrder",
        "parameters" : [ {
          "name" : "restaurantId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/restaurants/{restaurantId}/fulfillment/orders/{orderId}/accept" : {
      "post" : {
        "tags" : [ "fulfillment-controller" ],
        "operationId" : "acceptOrder",
        "parameters" : [ {
          "name" : "restaurantId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AcceptOrderRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/restaurants/{restaurantId}/fulfillment/orders" : {
      "get" : {
        "tags" : [ "fulfillment-controller" ],
        "operationId" : "getRestaurantOrders",
        "parameters" : [ {
          "name" : "restaurantId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListRestaurantOrder"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/restaurants/{restaurantId}/fulfillment/orders/history" : {
      "get" : {
        "tags" : [ "fulfillment-controller" ],
        "operationId" : "getHistoricalRestaurantOrders",
        "parameters" : [ {
          "name" : "restaurantId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "from",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "to",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "format" : "date-time"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePageResponseDtoRestaurantOrder"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/restaurants/{restaurantId}/fulfillment/orders/active" : {
      "get" : {
        "tags" : [ "fulfillment-controller" ],
        "operationId" : "getActiveRestaurantOrders",
        "parameters" : [ {
          "name" : "restaurantId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListRestaurantOrder"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ApiResponseVoid" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "object"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "PartialRefundRequestDto" : {
        "required" : [ "amount" ],
        "type" : "object",
        "properties" : {
          "amount" : {
            "minimum" : 0.01,
            "exclusiveMinimum" : false,
            "type" : "number"
          },
          "reason" : {
            "type" : "string"
          },
          "items" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "AcceptOrderRequest" : {
        "type" : "object",
        "properties" : {
          "additionalPrepTime" : {
            "type" : "integer",
            "format" : "int32"
          },
          "delayReason" : {
            "type" : "string"
          }
        }
      },
...
      "ApiResponseListRestaurantOrder" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RestaurantOrder"
            }
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "RestaurantOrder" : {
        "required" : [ "createdAt", "deliveryStatus", "orderId", "restaurantId", "status" ],
        "type" : "object",
        "properties" : {
          "orderId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "restaurantId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "CREATED", "PENDING_ACCEPTANCE", "AWAITING_DELAY_APPROVAL", "ACCEPTED", "PREPARING", "READY_FOR_PICKUP", "HANDED_OVER", "CANCELLED", "CANCELLED_BY_RESTAURANT", "CANCELLED_BY_PLATFORM", "DELIVERY_FAILED" ]
          },
          "deliveryStatus" : {
            "type" : "string",
            "enum" : [ "PENDING", "SEARCHING_FOR_DRIVER", "MANUAL_INTERVENTION_REQUIRED", "ASSIGNED", "AT_RESTAURANT", "OUT_FOR_DELIVERY", "DELIVERED", "CANCELLED", "FAILED" ]
          },
          "paymentStatus" : {
            "type" : "string",
            "enum" : [ "INITIATED", "SUCCESS", "FAILED", "PARTIALLY_REFUNDED", "REFUNDED", "REFUND_PENDING", "REFUND_FAILED" ]
          },
          "version" : {
            "type" : "integer",
            "format" : "int32"
          },
          "prepTime" : {
            "type" : "integer",
            "format" : "int32"
          },
          "additionalPrepTime" : {
            "type" : "integer",
            "format" : "int32"
          },
          "estimatedCompletionTime" : {
            "type" : "integer",
            "format" : "int64"
          },
          "deliveryLat" : {
            "type" : "number",
            "format" : "double"
          },
          "deliveryLng" : {
            "type" : "number",
            "format" : "double"
          },
          "deliveryAddress" : {
            "type" : "string"
          },
          "dispatchCityId" : {
            "type" : "string"
          },
          "fleetSearchRadiusKm" : {
            "type" : "number",
            "format" : "double"
          },
          "pickupOtp" : {
            "type" : "string"
          },
          "deliveryOtp" : {
            "type" : "string"
          },
          "deliveryExecutiveId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "customerId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "paymentMethod" : {
            "type" : "string",
            "enum" : [ "CARD", "UPI", "WALLET" ]
          },
          "customerName" : {
            "type" : "string"
          },
          "foodCost" : {
            "type" : "number"
          },
          "restaurantPlatformFee" : {
            "type" : "number"
          },
          "restaurantDeliveryContribution" : {
            "type" : "number"
          },
          "platformBonus" : {
            "type" : "number"
          },
          "restaurantPayout" : {
            "type" : "number"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "deliveryExecutiveName" : {
            "type" : "string"
          },
          "total" : {
            "type" : "number"
          },
          "items" : {
            "type" : "object"
          }
        }
      },
      "ApiResponsePageResponseDtoRestaurantOrder" : {
        "required" : [ "message", "success", "timestamp" ],
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean"
          },
          "message" : {
            "type" : "string"
          },
          "errorCode" : {
            "type" : "string"
          },
          "data" : {
            "$ref" : "#/components/schemas/PageResponseDtoRestaurantOrder"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PageResponseDtoRestaurantOrder" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/RestaurantOrder"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  }
}